pyproject.toml
[tool.poetry]
name = "logger"
version = "0.1.0"
description = ""
authors = ["phantie <funtom1605@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencensus-ext-azure = "^1.1.13"
opencensus-ext-logging = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
